refactor: Markers 하위 폴더 코드 분리
[#809]
