+++ frontend/src/components/google-maps/marker/CityClusteredMarkersContainer/index.ts
@@ -1,0 +1,3 @@
+import CityClusteredMarkersContainer from './CityClusteredMarkersContainer';
+
+export default CityClusteredMarkersContainer;
