+++ frontend/src/components/google-maps/marker/ServerClusteredMarkersContainer/ServerClusteredMarkersContainer.tsx
