+++ frontend/src/components/google-maps/marker/ServerClusteredMarkersContainer/index.ts
@@ -1,0 +1,3 @@
+import ServerClusteredMarkersContainer from './ServerClusteredMarkersContainer';
+
+export default ServerClusteredMarkersContainer;
